Move `EventLoopHandler` to dedicated module
