+++ crates/fj-window/src/lib.rs
@@ -15,3 +15,5 @@
 
 pub mod run;
 pub mod window;
+
+mod event_loop_handler;
